refactor(tests): removed redundant typings and adjusted tests to use the correct ones
