--- src/qiboconnection/typings/saved_experiment.py
@@ -1,60 +1,0 @@
-# Copyright 2023 Qilimanjaro Quantum Tech
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-""" SavedExperiments typing classes """
-from abc import ABC
-from dataclasses import dataclass
-from datetime import datetime
-
-
-@dataclass
-class _SavedExperimentBodyBase(ABC):
-    """Base structure for a SavedExperiment body of any rest interaction"""
-
-    name: str
-    experiment: str
-    device_id: int
-    user_id: int
-    description: str
-    qililab_version: str
-
-
-@dataclass
-class _FullSavedExperimentBodyBase(_SavedExperimentBodyBase):
-    """Base structure for a SavedExperiment body of a rest interaction that includes results"""
-
-    results: str
-
-
-@dataclass
-class SavedExperimentRequest(_FullSavedExperimentBodyBase):
-    """Class for accommodating SavedExperiments web requests."""
-
-    favourite: bool = False
-
-
-@dataclass
-class SavedExperimentResponse(_FullSavedExperimentBodyBase):
-    """Class for accommodating SavedExperiments web responses."""
-
-    saved_experiment_id: int
-    created_at: datetime
-
-
-@dataclass
-class SavedExperimentListingItemResponse(_SavedExperimentBodyBase):
-    """Class for accommodating SavedExperimentListings web responses"""
-
-    id: int
-    created_at: datetime
